Enable Java-behavior for late cause initialisation
Java deals differently for a cause of null and an unset cause. If the
cause is initialized with null subsequent calls to initCause() will
fail. If the cause is unset a subsequent call to initCause() works.

This solution introduces a "flag-exception" to indicate that no cause
has been set and thus the cause gets only initialized if the cause has
been explicitly set by the API user.

Fixes #70.
Closes #19.