Issue #21: Introduce DebugInformation
Introducing a more generic way to provide additional information
for debugging upon assertion failures. The DebugInformation is
key-value based and can contain actual any information.

In addition to this extended the ValueWrapper (now renamed to
ValueDescriptor as it *describes* the value) so that it does some
effort to serialize objects applying some strategies. This way we
could also provide images as for example for UI tests.

The current state neither includes documentation, more tests and
misses the commonly used exception for all exceptions provided
inside here.

This approach also addresses issues:

* #20 support for exception creation strategies
* #4 detach AssertionError from exceptions here
* #9 partially, as you the debug information provided here should
  eventually merge into a standard test report.