Change MultipleFailuresError to support Throwables
As individual failures had to be childs of AssertionError this made
MultipleFailuresError unsuitable for JUnit4 tests using ErrorCollector
which can register multiple failures of type Throwable.

Also adds a null check for throwables in getMessage(). Each added throwable
is -in essense- mapped to an instance created with
org.junit.platform.engine.TestExecutionResult.failed() which is very
clearly documented to accept null values. In that case the message
method would crash.

Resolves: #30, junit-team/junit5#659